Use float instead of int for Knob values

+++ tests/Widgets/Knob.cpp
@@ -36,9 +36,9 @@ TEST_CASE("[Knob]")
     SECTION("Signals")
     {
         REQUIRE_NOTHROW(knob->connect("ValueChanged", [](){}));
-        REQUIRE_NOTHROW(knob->connect("ValueChanged", [](int){}));
+        REQUIRE_NOTHROW(knob->connect("ValueChanged", [](float){}));
         REQUIRE_NOTHROW(knob->connect("ValueChanged", [](tgui::Widget::Ptr, std::string){}));
-        REQUIRE_NOTHROW(knob->connect("ValueChanged", [](tgui::Widget::Ptr, std::string, int){}));
+        REQUIRE_NOTHROW(knob->connect("ValueChanged", [](tgui::Widget::Ptr, std::string, float){}));
     }
 
     SECTION("WidgetType")
